car_market/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5  #python sql köprüsü
Flask-Login==0.6.3   #giriş yapma, beni hatırla
Flask-WTF==1.1.1      #ilan verme, kayıt olma güvenliği
Flask-Migrate==4.0.4   #veritabanı değişikliklleri
psycopg2-binary==2.9.7    #PostgreSQL veritabanı bağlayıcı
Pillow==10.4.0          #resim işlemleri
python-dotenv==1.0.0     #gizli anahtar ve veritabanı şifreleri .env de
WTForms==3.0.1
email-validator==2.1.0   #emailler doğru formatta mı